refactor: FileCriteria component to FileCriteriaForm
