+++ tests/component/sections/dataset/dataset-files/file-criteria-form/FileCriteriaFilterByAccess.spec.tsx
@@ -3,8 +3,8 @@ import {
   FileCriteria
 } from '../../../../../../src/files/domain/models/FileCriteria'
 import { FilesCountInfoMother } from '../../../../files/domain/models/FilesCountInfoMother'
-import styles from '../../../../../../src/sections/dataset/dataset-files/file-criteria-controls/FileCriteriaControls.module.scss'
-import { FileCriteriaFilterByAccess } from '../../../../../../src/sections/dataset/dataset-files/file-criteria-controls/FileCriteriaFilterByAccess'
+import styles from '../../../../../../src/sections/dataset/dataset-files/file-criteria-form/FileCriteriaControls.module.scss'
+import { FileCriteriaFilterByAccess } from '../../../../../../src/sections/dataset/dataset-files/file-criteria-form/FileCriteriaFilterByAccess'
 
 const defaultCriteria = new FileCriteria()
 const filesCountInfo = FilesCountInfoMother.create({
